+++ System/Collections/List.lua
@@ -8,8 +8,8 @@
 -- Author       :   kurapica125@outlook.com                                  --
 -- URL          :   http://github.com/kurapica/PLoop                         --
 -- Create Date  :   2016/02/28                                               --
--- Update Date  :   2019/12/02                                               --
--- Version      :   1.2.2                                                    --
+-- Update Date  :   2020/02/19                                               --
+-- Version      :   1.2.3                                                    --
 --===========================================================================--
 
 PLoop(function(_ENV)
@@ -320,6 +320,8 @@ PLoop(function(_ENV)
             tremove             = table.remove,
             getobjectclass      = Class.GetObjectClass,
             issubtype           = Class.IsSubType,
+
+            ListStreamWorker, IIndexedList, ICountable
         }
 
         -----------------------------------------------------------
@@ -332,13 +334,14 @@ PLoop(function(_ENV)
             export { Context }
 
             getIdleworkers      = function()
-                local context   = Context.Current
-                return context and context[ListStreamWorker]
+                -- Disable the recycle in multi os thread
+                -- local context   = Context.Current
+                -- return context and context[ListStreamWorker]
             end
 
             rycIdleworkers      = function(worker)
-                local context   = Context.Current
-                if context then context[ListStreamWorker] = worker end
+                -- local context   = Context.Current
+                -- if context then context[ListStreamWorker] = worker end
             end
         else
             -- Keep idle workers for re-usage
@@ -563,12 +566,14 @@ PLoop(function(_ENV)
         --- Map the items to other datas
         __Arguments__{ Callable }
         function Map(self, func)
+            if self[FLD_MAPACTITON] then return ListStreamWorker(self):Map(func) end
             self[FLD_MAPACTITON] = func
             return self
         end
 
         __Arguments__{ String }
         function Map(self, feature)
+            if self[FLD_MAPACTITON] then return ListStreamWorker(self):Map(feature) end
             self[FLD_MAPACTITON] = function(item)
                 if type(item) == "table" then
                     return item[feature]
@@ -579,6 +584,7 @@ PLoop(function(_ENV)
 
         __Arguments__{ Callable }
         function Filter(self, func)
+            if self[FLD_FILTERACTN] then return ListStreamWorker(self):Filter(func) end
             self[FLD_FILTERACTN] = func
             return self
         end
@@ -586,6 +592,7 @@ PLoop(function(_ENV)
         --- Used to filter the items with a check function
         __Arguments__{ String, System.Any/nil }
         function Filter(self, feature, value)
+            if self[FLD_FILTERACTN] then return ListStreamWorker(self):Filter(feature, value) end
             self[FLD_FILTERACTN] = value ~= nil and function(item)
                 if type(item) == "table" then
                     return item[feature] == value
@@ -605,6 +612,7 @@ PLoop(function(_ENV)
         --- Used to select items with ranged index
         __Arguments__{ Integer/1, Integer/-1, Integer/1 }
         function Range(self, start, stop, step)
+            if self[FLD_RANGESTART] then return ListStreamWorker(self):Range(start, stop, step) end
             self[FLD_RANGESTART], self[FLD_RANGESTOP], self[FLD_RANGESTEP] = start, stop, step
             return self
         end
@@ -613,12 +621,12 @@ PLoop(function(_ENV)
         --                      Constructor                      --
         -----------------------------------------------------------
         __Arguments__{ IList }
-        function ListStreamWorker(self, list)
+        function __ctor(self, list)
             self[FLD_TARGETLIST] = list
         end
 
         __Arguments__{ Callable, System.Any/nil, System.Any/nil }
-        function ListStreamWorker(self, iter, obj, idx)
+        function __ctor(self, iter, obj, idx)
             self[FLD_TARGETITER] = iter
             self[FLD_ITEROBJECT] = obj
             self[FLD_ITERINDEX] = idx
@@ -644,8 +652,6 @@ PLoop(function(_ENV)
             end
             return worker
         end
-
-        export{ ListStreamWorker, IIndexedList, ICountable }
     end)
 
     __Sealed__()